MATH-AI/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
opencv-python==4.8.1.78
requests==2.31.0
openai==1.3.0
anthropic>=0.18.0
google-generativeai>=0.3.0
websockets==12.0
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
numpy>=1.26.0
